pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "webhook-relay"
version = "0.1.0"
description = "A system for relaying webhooks from external sources to internal services"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "fastapi>=0.95.0",
    "uvicorn>=0.20.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "loguru>=0.7.0",
    "click>=8.1.0",
    "aiohttp>=3.8.0",
    "prometheus-client>=0.16.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
gcp = ["google-cloud-pubsub>=2.17.0"]
aws = ["boto3>=1.28.0"]
dev = [
    "pytest>=7.3.1",
    "black>=23.3.0",
    "isort>=5.12.0",
    "mypy>=1.3.0",
    "flake8>=6.0.0",
]
all = ["webhook-relay[gcp,aws,dev]"]

[project.scripts]
webhook-relay-collector = "webhook_relay.collector.app:cli"
webhook-relay-forwarder = "webhook_relay.forwarder.app:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/webhook_relay"]

[tool.isort]
profile = "black"
line_length = 88

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = "google.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "boto3.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "prometheus_client.*"
ignore_missing_imports = true